refactor: Updated the namespace of utilities
