+++ tests/dummy/app/utils/components/widgets/widget-2.js
