+++ tests/dummy/app/utils/components/widgets/widget-3.js
